Optimize EmojiMemoriGameView Code
