+++ P1/EmojiMemoryGame.swift
@@ -14,7 +14,7 @@ class EmojiMemoryGame {
     private static let emojis  = ["👻","🙋‍♀️","🥹","😻","😇","🥰","😎","😆","🤓","😋","😇","😻"]
     
     private static func createMemoryGame() -> MemoryGame<String> {
-        MemoryGame(numberOfPairsOfCards: 4) {pairIndex in
+        MemoryGame(numberOfPairsOfCards: 8) {pairIndex in
             if emojis.indices.contains(pairIndex){
                 return emojis[pairIndex]
