+++ P1/MemoryGame.swift
@@ -26,9 +26,9 @@ struct MemoryGame<CardContent>{
     func choose(_ card:Card){
  
     }
-    
+
     struct Card{
-        var isFaceUp = false
+        var isFaceUp = true
         var isMatched = false
         var content : CardContent
     }
