+++ P1/P1App.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct P1App: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            EmojiMemoryGameView()
         }
     }
 }
